requirements.txt
# Generated file, do not edit
Flask-Classful>=0.14.2
python-arango>=4.4.0
marshmallow>=2.19.0,<3
pytest-runner
Sphinx>=1.8.0
sphinx-rtd-theme
setuptools
wheel
pytest>=4.2.0
pytest-cov>=2.6.0
pytest-flake8
